requirements.txt
configparser=6.0.0
requests
tk
Adafruit-Blinka
adafruit-circuitpython-ads1x15
keyboard
Pillow


setuptools~=66.1.1
pyserial~=3.5